engine/requirements.txt
pyyaml>=6.0
jsonschema>=4.20.0
requests>=2.31.0
jinja2>=3.1.0
fastmcp>=0.2.0
